concept separated to another component, problem with yes and yes and buttons v frustrating
