+++ peerdea-native/components/Concept.js
@@ -1,0 +1,296 @@
+import React, { Component } from 'react';
+import { Button,
+          Image,
+          View,
+          StyleSheet,
+          Text,
+          TouchableOpacity,
+          TouchableHighlight,
+          TextInput,
+          Alert,
+          ScrollView,
+          Modal } from 'react-native';
+// because we're using mangaged apps version of expo (and not bare version):
+import { ImagePicker,
+         Permissions,
+         Camera } from 'expo';
+import { Buffer } from 'buffer';
+import Collapsible from 'react-native-collapsible';
+import ImageCarousel from 'react-native-image-carousel';
+
+// PICK UP HERE
+//TODO: change infrastructure of this file to make state hold multple values of what a concet is
+
+export default class Concept extends React.Component {
+
+constructor(props) {
+    super(props);
+    this.state = {
+      concept: props.concept,
+      modalVisible: false,
+      modalErrorMessage: '',
+      iLike: '',
+      iWish: '',
+    };
+    this.updateConcept=this.updateConcept.bind(this);
+    this._yes = this._yes.bind(this);
+    this._yesAnd = this._yesAnd.bind(this);
+    this._changeCollapse=this._changeCollapse.bind(this);
+    this._sendFeedback = this._sendFeedback.bind(this);
+}
+
+
+
+  componentDidMount() {
+    //this.updateConcept();
+  };
+
+  async updateConcept() {
+    const res = await fetch('http://104.40.20.156/api/getConceptByID?id=' + this.state.concept._id, {method: 'GET'});
+    const resJson = await res.json();
+    const conceptNew = resJson.data;
+    conceptNew.isCollapsed = true;
+    console.log("updatin");
+    console.log(conceptNew);
+
+    await this.setState({concept: conceptNew});
+    console.log(this.state.concept);
+  }
+
+  render() {
+      var yesAndViews = []
+      for (j = 0; j < this.state.concept.yesand.length; j++){
+          const yesandText = this.state.concept.yesand[j];
+          yesAndViews.push(
+                  <Text key = {j}> {yesandText} </Text>
+          )
+      }
+      const yesAnds = yesAndViews;
+      var images = [];
+      for (imageI = 0; imageI < this.state.concept.media.length; imageI++){
+          const buff = new Buffer(this.state.concept.media[imageI].data);
+          const base64data = buff.toString('base64');
+          const uriString = `data:image/gif;base64,${base64data}`;
+          images.push(uriString);
+      }
+      const finalImages = images;
+    return (
+      <View key = {i} style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+				<View style={{flex: 1}}>
+          <ImageCarousel
+            renderContent={(idx: number) => (
+                <Image
+                  style={StyleSheet.absoluteFill}
+                  resizeMode="contain"
+                  source={{uri: finalImages[idx]}}
+                />
+              )}>
+            {finalImages.map(url => (
+              <Image
+                style={{ width: 200, height: 200 }}
+                key={url}
+                source={{uri: url}}
+                resizeMode="contain"
+              />
+            ))}
+           </ImageCarousel>
+         </View>
+          <Text> {this.state.concept.name} </Text>
+          <Text> {this.state.concept.description} </Text>
+          <TouchableOpacity style={styles.btn}  onPress = {async () => { await this._yes();}}>
+             <Text>Yes {this.state.concept.yes}</Text>
+             <Image source={require('../assets/images/heart.png')}  style={{width: 20, height: 20}}/>
+          </TouchableOpacity>
+          <TouchableOpacity style={styles.btn} onPress = {() => { this._yesAnd();}}>
+              <Text>Yes And</Text>
+              <Image source={require('../assets/images/heart.png')}  style={{ width: 20, height: 20}}/>
+          </TouchableOpacity>
+          {this.state.concept.isCollapsed &&
+              <Button title="Expand" onPress={() => this._changeCollapse(false)}/>}
+          <Collapsible collapsed={this.state.concept.isCollapsed}>
+              <Button title="Collapse" onPress={() => this._changeCollapse(true)}/>
+              {yesAnds}
+          </Collapsible>
+          <Text>  {"\n\n"} </Text>
+
+      {this.state.modalVisible && <Modal
+          animationType="slide"
+          transparent={false}
+          visible={true}
+          onRequestClose={() => {
+            console.log('Modal is closed'); //onRequestClose is a required parameter of the Modal component
+          }}>
+          <View style={{marginTop: 100, alignItems:'center'}}>
+              <Text style={{color: 'red', fontWeight: 'bold'}}> {this.state.modalErrorMessage} </Text>
+              <Text style={styles.getStartedText}>Enter your feedback:</Text>
+              <View style={{flexDirection: 'row'}}>
+                <TextInput
+                  style={{height: 40, flex: 0.5, borderColor: 'gray', borderWidth: 1}}
+                  onChangeText={(text) => this.setState({iLike: text})}
+                  placeholder="I like..."
+                />
+              </View>
+              <View style={{flexDirection: 'row'}}>
+                <TextInput
+                  style={{height: 40, flex: 0.5, borderColor: 'gray', borderWidth: 1}}
+                  onChangeText={(text) => this.setState({iWish: text})}
+                  placeholder="I wish..."
+                />
+              </View>
+              <Button raised
+                onPress={() =>
+                    {
+                      this._sendFeedback(this.state.iLike, this.state.iWish);
+
+                    }
+                }
+                title="Submit feedback"
+                color="#841584"
+                accessibilityLabel="Submit feedback"
+              />
+              <TouchableHighlight
+                onPress={() => {
+                  //if user decides to cancel, close modal and reset state for iLike and iWish, as user has decided to cancel submission of feedback
+                  this.setState({ modalErrorMessage: '',
+                                  iLike: '',
+                                  iWish: '',
+                                  modalVisible:false,});
+                }}>
+                <Text>Cancel</Text>
+              </TouchableHighlight>
+            </View>
+        </Modal>
+
+      }
+      </View>
+    );
+  }
+
+
+
+  _sendFeedback = async (iLike, iWish) => {
+
+      var update = false;
+
+     if (iLike.length > 0) {
+       if (!iLike.startsWith("I like")){
+            this.setState({ modalErrorMessage: 'Feedback in first spot must start with I like'});
+            return;
+
+       }else{
+           let data = {
+           method: 'POST',
+           credentials: 'same-origin',
+           mode: 'same-origin',
+           body: JSON.stringify({
+             id: this.state.concept._id,
+             text: iLike,
+           }),
+           headers: {
+             'Accept':       'application/json',
+             'Content-Type': 'application/json',
+             // 'X-CSRFToken':  cookie.load('csrftoken')
+             }
+            }
+           const res = await fetch('http://104.40.20.156/api/yesand', data)
+            updated = true;
+        }//if testing locally, insert your machines IP along with the port number set in server.js e.g. :3000
+
+     }
+
+     if (iWish.length > 0) {
+        if (!iWish.startsWith("I wish")){
+            this.setState({ modalErrorMessage: 'Feedback in second spot must start with I wish'});
+            return;
+       }else{
+           let data2 = {
+           method: 'POST',
+           credentials: 'same-origin',
+           mode: 'same-origin',
+           body: JSON.stringify({
+             id: this.state.concept._id,
+             text: iWish,
+           }),
+           headers: {
+             'Accept':       'application/json',
+             'Content-Type': 'application/json',
+             // 'X-CSRFToken':  cookie.load('csrftoken')
+             }
+            }
+
+           const res = await fetch('http://104.40.20.156/api/yesand', data2);
+            updated = true;
+         }//if testing locally, insert your machines IP along with the port number set in server.js e.g. :3000
+     }
+     if (updated){
+        this.updateConcept();
+        //after user send feedback, close modal and reset state for iLike and iWish, as user has decided to cancel submission of feedback
+        this.setState({ iLike: '',
+                      modalErrorMessage: '',
+                      iWish: '',
+                      modalVisible:false,});
+     }
+
+    // else
+    //   console.log('User did not input anything');
+
+
+   }
+
+  _changeCollapse = async (val) => {
+        //console.log(val)
+        conceptNew = this.state.concept;
+        conceptNew.isCollapsed = val;
+        this.setState({concepts: conceptNew});
+  }
+
+  _yes = async () => {
+     let data = {
+       method: 'POST',
+       credentials: 'same-origin',
+       mode: 'same-origin',
+       body: JSON.stringify({
+         id: this.state.concept._id
+       }),
+       headers: {
+         'Accept':       'application/json',
+         'Content-Type': 'application/json',
+         // 'X-CSRFToken':  cookie.load('csrftoken')
+         }
+        }
+       const res = await fetch('http://104.40.20.156/api/yes', data)
+       this.updateConcept();
+  }
+
+// id here is the concept id
+  _yesAnd = async (id) => {
+    // console.log("we are in _yesand");
+    // console.log("ID IS " + id);
+    //interestingly, the order in which you update state variable in setState matters
+    //thisConceptID must be updated before modalVisible
+    this.setState({modalVisible: true});
+  }
+}
+
+
+const styles = StyleSheet.create({
+  container: {
+    paddingTop: 30,
+    paddingBottom: 30,
+    flex: 1,
+  },
+  getStartedText: {
+    fontSize: 17,
+    color: 'rgba(96,100,109, 1)',
+    lineHeight: 24,
+    textAlign: 'center',
+  },
+  contentContainer: {
+    paddingTop: 30,
+  },
+   absoluteView: {
+        flex: 2,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+});
